+++ Soggins/Artifacts/BotchTrackerArtifact.cs
@@ -1,0 +1,54 @@
+﻿using CobaltCoreModding.Definitions.ExternalItems;
+using CobaltCoreModding.Definitions.ModContactPoints;
+using Shockah.Shared;
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+namespace Shockah.Soggins;
+
+[ArtifactMeta(unremovable = true)]
+public sealed class BotchTrackerArtifact : Artifact, IRegisterableArtifact
+{
+	private static ModEntry Instance => ModEntry.Instance;
+
+	private static ExternalSprite Sprite = null!;
+
+	public int Botches;
+
+	public void RegisterArt(ISpriteRegistry registry)
+	{
+		Sprite = registry.RegisterArtOrThrow(
+			id: $"{GetType().Namespace}.Artifact.BotchTracker",
+			file: new FileInfo(Path.Combine(Instance.ModRootFolder!.FullName, "assets", "Artifact", "BotchTracker.png"))
+		);
+	}
+
+	public void RegisterArtifact(IArtifactRegistry registry)
+	{
+		ExternalArtifact artifact = new(
+			globalName: $"{GetType().Namespace}.Artifact.BotchTracker",
+			artifactType: GetType(),
+			sprite: Sprite,
+			ownerDeck: Instance.SogginsDeck
+		);
+		artifact.AddLocalisation(I18n.BotchTrackerArtifactName.ToUpper(), I18n.BotchTrackerArtifactDescription);
+		registry.RegisterArtifact(artifact);
+	}
+
+	public override int? GetDisplayNumber(State s)
+		=> s.route is Combat ? Botches : null;
+
+	public override List<Tooltip>? GetExtraTooltips()
+	{
+		var tooltips = base.GetExtraTooltips() ?? [];
+		tooltips.Add(Instance.Api.GetSmugTooltip());
+		return tooltips;
+	}
+
+	public override void OnCombatStart(State state, Combat combat)
+	{
+		base.OnCombatStart(state, combat);
+		Botches = 0;
+	}
+}