+++ Soggins/docs/release-notes.md
@@ -5,6 +5,7 @@
 ## Upcoming release
 
 * Changed the Stop It! card, effectively merging it with Zen.
+* Replaced the Botches fake status with a dedicated botch tracker artifact.
 * Fixed Smug-related dialogue not being properly serialized, which normally wouldn't really be noticeable, unless you're playing with the Combat QoL mod.
 * Apology-generating cards now use their own unique action.